backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# Redis & Caching
redis==5.0.1
hiredis==2.2.3

# Async Task Processing
celery==5.3.4
flower==2.0.1

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
cohere==4.37
tiktoken==0.5.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Text Processing & NLP
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# Observability
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
ruff==0.1.6
mypy==1.7.1

